requirements.txt
asgiref==3.8.1
async-generator==1.10
attrs==24.2.0
black==24.8.0
build==1.2.1
certifi==2024.8.30
cffi==1.17.0
click==8.1.7
colorama==0.4.6
coverage==7.6.1
Django==5.1
django-bootstrap3==24.2
django-crum==0.7.9
exceptiongroup==1.2.2
h11==0.14.0
idna==3.8
iniconfig==2.0.0
isort==5.13.2
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
pip-review==1.3.0
pip-tools==7.4.1
platformdirs==4.2.2
pluggy==1.5.0
pycparser==2.22
pyproject_hooks==1.1.0
PySocks==1.7.1
pytest==8.3.2
pytest-django==4.8.0
pytz==2024.1
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tomli==2.0.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
